\name{is.irt.pars}
\alias{is.irt.pars}
\alias{is.sep.pars}
\alias{is.poly.mod}
\alias{is.irt.prob}
\alias{is.link}
\title{Identify Class}
\description{
These functions test whether a given object is of a specified class
}
\usage{
is.irt.pars(x)
is.sep.pars(x)
is.poly.mod(x)
is.irt.prob(x)
is.link(x)
}

\arguments{
  \item{x}{an \code{R} object}
}

\value{
Returns a logical value
}

\author{ Jonathan P. Weeks \email{weeksjp@gmail.com} }

\seealso{\code{\linkS4class{irt.pars}}, \code{\linkS4class{sep.pars}}, \code{\linkS4class{poly.mod}},
\code{\linkS4class{irt.prob}}, \code{\linkS4class{link}} }

\keyword{ misc }
