\name{list.poly-class}
\docType{class}
\alias{list.poly-class}
\alias{list.dat-class}
\alias{list.mat-class}
\alias{list.num-class}
\alias{num.null-class}
\alias{list.null-class}
\alias{pars.null-class}
\title{Class Combinations}
\description{These classes are combinations of classes}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}

\details{
  \describe{
    \item{list.poly:}{combines class \code{list} and class \code{\linkS4class{poly.mod}} }

    \item{list.dat:}{combines class \code{list} and class \code{data.frame} }

    \item{list.mat:}{combines class \code{list} and class \code{matrix} }

    \item{list.num:}{combines class \code{list} and class \code{numeric} }

    \item{num.null:}{combines class \code{numeric} and class \code{NULL} }

    \item{list.null:}{combines class \code{list} and class \code{NULL} }

    \item{pars.null:}{combines class \code{irt.pars} and class \code{NULL} }
  }
}

\author{ Jonathan P. Weeks \email{weeksjp@gmail.com} }

\keyword{classes}
