\name{md.mixed}
\alias{md.mixed}
\docType{data}
\title{Multidimensional Mixed-Format Tests}
\description{
This dataset includes multidimensional 3PL and multidimensional generalized partial
credit model item parameters for two groups modeled with four dimensions.  The item
parameters were estimated based on item responses provided with BMIRT (Yao, 2008).
}
\usage{md.mixed}
\format{
A list of length two. There are two matrices in the first list element that correspond
to form.x and form.y tests respectively.  The second list element identifies the 
number of response categories for each item.  All of the items are common items.  There
is no location parameter for the polytomous items.
}

\source{
  Yao, L. (2008). 
  BMIRT: Bayesian multivariate item response theory [Computer Program]. 
  Monterey, CA: CTB/McGraw-Hill.
}

\keyword{datasets}
